requirements.txt
moviepy==1.0.3
Pillow==10.3.0
numpy>=1.20.0
